Revert "add singlton calculator class"
